Restore board listing methods
