+++ test/integration.tests/Dotnet.AzureDevOps.Core.ProjectSettingsIntegatioTests/DotnetAzureDevOpsProjectSettingsIntegrationTests.cs
@@ -1,0 +1,58 @@
+using System.Diagnostics.CodeAnalysis;
+using Dotnet.AzureDevOps.Core.Boards;
+using Dotnet.AzureDevOps.Core.ProjectSettings;
+using Dotnet.AzureDevOps.Tests.Common;
+using Microsoft.TeamFoundation.Core.WebApi;
+using Microsoft.TeamFoundation.Work.WebApi;
+
+namespace Dotnet.AzureDevOps.Core.ProjectSettingsIntegatioTests
+{
+    [ExcludeFromCodeCoverage]
+    public class DotnetAzureDevOpsProjectSettingsIntegrationTests : IAsyncLifetime
+    {
+        private readonly AzureDevOpsConfiguration _azureDevOpsConfiguration;
+        private readonly ProjectSettingsClient _projectSettingsClient;
+        private readonly WorkItemsClient _workItemsClient;
+
+        public DotnetAzureDevOpsProjectSettingsIntegrationTests()
+        {
+            _azureDevOpsConfiguration = new AzureDevOpsConfiguration();
+            _projectSettingsClient = new ProjectSettingsClient(
+                _azureDevOpsConfiguration.OrganisationUrl,
+                _azureDevOpsConfiguration.ProjectName,
+                _azureDevOpsConfiguration.PersonalAccessToken);
+            _workItemsClient = new WorkItemsClient(
+                _azureDevOpsConfiguration.OrganisationUrl,
+                _azureDevOpsConfiguration.ProjectName,
+                _azureDevOpsConfiguration.PersonalAccessToken);
+        }
+
+        [Fact]
+        public async Task TeamAndBoardConfiguration_SucceedsAsync()
+        {
+            string testTeamName = "Dotnet.McpIntegrationTest Team";
+            var teamContext = new TeamContext(_azureDevOpsConfiguration.ProjectName, testTeamName);
+            string boardName = $"{_azureDevOpsConfiguration.ProjectName} Team";
+            await _projectSettingsClient.CreateTeamAsync(testTeamName, "description1");
+            await _projectSettingsClient.UpdateTeamDescriptionAsync(testTeamName, "description2");
+            List<BoardReference> boardReferenceList = await _workItemsClient.ListBoardsAsync(teamContext, boardName);
+            List<TeamSettingsIteration> iterations = await _workItemsClient.GetTeamIterationsAsync(teamContext, "");
+
+            IReadOnlyList<BoardColumn> cols = await _workItemsClient.ListBoardColumnsAsync(teamContext, boardReferenceList[0].Id, testTeamName);
+            await _projectSettingsClient.DeleteTeamAsync(await _projectSettingsClient.GetTeamIdAsync(testTeamName));
+
+            Assert.NotNull(cols);
+
+            IReadOnlyList<TeamSettingsIteration> iterationList = await _workItemsClient.ListIterationsAsync(teamContext, "current", _azureDevOpsConfiguration.ProjectName);
+            Assert.NotNull(iterations);
+            Assert.NotNull(iterationList);
+
+            TeamFieldValues areas = await _workItemsClient.ListAreasAsync(teamContext);
+            Assert.NotNull(areas);
+        }
+
+        public Task InitializeAsync() => Task.CompletedTask;
+
+        public Task DisposeAsync() => Task.CompletedTask;
+    }
+}
