+++ test/integration.tests/Dotnet.AzureDevOps.Core.ProjectSettingsIntegatioTests/xunit.runner.json
@@ -1,0 +1,5 @@
+{
+    "parallelizeAssembly": true,
+    "parallelizeTestCollections": false,
+    "maxParallelThreads": 1
+}
